requirements.txt
transformers
requests
python-dotenv
Flask
gunicorn
torch==2.5.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu